# FishingJournalAPI/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FishingJournalAPI",
    "version": "v1"
  },
  "paths": {
...
    "/JournalEntry": {
      "get": {
        "tags": [
          "JournalEntry"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/JournalEntry/{id}": {
      "get": {
        "tags": [
          "JournalEntry"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/JournalEntry/{username}": {
      "get": {
        "tags": [
          "JournalEntry"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter valid token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}